project

+++ backend/controllers/product.controller.js
@@ -15,13 +15,17 @@ export const getProducts = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     console.log("id : ", req.params)
     const { id } = req.params
+    
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: "Invalid product ID"})
+    }
 
     try {
         await Product.findByIdAndDelete(id)
         res.status(200).json({ success: true, message: "Product deleted." })
     } catch (error) {
         console.log("error fetching : ", error.message)
-        res.status(404).json({ status: false, message: "Product not found."})
+        res.status(500).json({ status: false, message: "Server error."})
     }
 
 }
